pyproject.toml
[tool.poetry]
name = "human-api"
version = "0.1.0"
description = "The Human API"
authors = ["Vlad Silviu Farcas <vlad@intuitionmachines.com, Harjyot Singh <harry@intuitionmachines.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
trinity = {git = "https://github.com/ethereum/trinity", rev = "eth2/departure"}
aiohttp = "*"
alabaster = "*"
appnope = "*"
argcomplete = "*"
asks = "*"
asn1crypto = "*"
async_generator = "*"
async_lru = "*"
async-timeout = "*"
asyncio-cancel-token = "*"
attrs = "*"
Babel = "*"
backcall = "*"
bitarray = "*"
blake2b-py = "*"
bloom-filter = "*"
boto3 = "*"
botocore = "*"
cached-property = "*"
cachetools = "*"
certifi = "*"
chardet = "*"
click = "*"
clickclick = "*"
cloudpickle = "*"
coincurve = "*"
connexion = "*"
cytoolz = "*"
decorator = "*"
docutils = "*"
eth-abi = "*"
eth-account = "*"
eth-bloom = "*"
eth-enr = "*"
eth-hash = "*"
eth-keyfile = "*"
eth-keys = "*"
eth-rlp = "*"
eth-typing = "*"
eth-utils = "*"
Flask = "*"
Flask-PyMongo = "*"
h11 = "*"
hexbytes = "*"
hmt-basemodels = "*"
hmt-escrow = "*"
idna = "*"
ifaddr = "*"
imagesize = "*"
inflection = "*"
ipfshttpclient = "*"
ipython_genutils = "*"
iso8601 = "*"
itsdangerous = "*"
jedi = "*"
Jinja2 = "*"
jmespath = "*"
jsonschema = "*"
lxml = "*"
MarkupSafe = "*"
marshmallow = "*"
multiaddr = "*"
multidict = "*"
mypy = "*"
mypy-extensions = "*"
netaddr = "*"
netdisco = "*"
netifaces = "*"
openapi-spec-validator = "*"
outcome = "*"
p2p = "*"
packaging = "*"
parsimonious = "*"
parso = "*"
pexpect = "*"
pickleshare = "*"
plyvel = "*"
prometheus-client = "*"
prompt-toolkit = "*"
protobuf = "*"
psutil = "*"
ptyprocess = "*"
py = "*"
py-solc-x = "*"
pycparser = "*"
pycryptodome = "*"
pydantic = "*"
pyethash = "*"
pyformance = "*"
Pygments = "*"
pymongo = "*"
pyparsing = "*"
pyrsistent = "*"
pysha3 = "*"
python-dateutil = "*"
python-dotenv = "*"
python-snappy = "*"
pytz = "*"
PyYAML = "*"
requests = "*"
retry = "*"
rusty-rlp = "*"
s3transfer = "*"
schematics = "*"
semantic-version = "*"
six = "*"
sniffio = "*"
snowballstemmer = "*"
sortedcontainers = "*"
sphinxcontrib-applehelp = "*"
sphinxcontrib-devhelp = "*"
sphinxcontrib-htmlhelp = "*"
sphinxcontrib-jsmath = "*"
sphinxcontrib-qthelp = "*"
sphinxcontrib-serializinghtml = "*"
SQLAlchemy = "*"
swagger-ui-bundle = "*"
termcolor = "*"
toolz = "*"
traitlets = "*"
trie = "2.0.0-alpha.2"
trio = "*"
trio-typing = "*"
typed-ast = "*"
typing-extensions = "*"
upnp-port-forward = "*"
upnpclient = "*"
urllib3 = "*"
uvloop = "*"
varint = "*"
wcwidth = "*"
web3 = "*"
websockets = "*"
Werkzeug = "*"
yarl = "*"
zeroconf = "*"
Flask-Testing = "*"
coverage = "*"
nose = "*"
pluggy = "*"
randomize = "*"
tox = "*"
tox-poetry = "^0.3.0"
yapf = "^0.30.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
